INA219 battery V+I on every uplink: frame fmt 0x02, 32 bytes (LOGBOOK r22)
batt_mA (i16, discharge positive, 0x7FFF = no INA219) appended at offset
30; all prior offsets unchanged. Selftest packer vector extended (passing
byte-exact on hardware); TTN decoder accepts fmt 0x01 and 0x02. CSV gains
a batt_mA column in the flash dump and SD file; flash-ring record now
carries the 32-byte frame (old records age out via checksum). INA219 init
probes 0x45 then 0x40. Battery model corrected to the fitted pack: 1S
Li-ion 2P (2x6600 mAh = 13.2 Ah, 4.2 V full / 3.0 V empty) replacing the
inherited 4S LiFePO4 constants. Live-verified: 4.084 V, +118 mA
discharging. Console batt line now prints V, mA and charge direction.

+++ CM4/EnviroNode_CM4/Core/WL55JC1/Src/envnode_log.c
@@ -36,8 +36,8 @@ typedef struct __attribute__((aligned(8))) {
   uint32_t epoch;
   uint8_t  frame[ENVLOG_FRAME_LEN];
   uint16_t sum;
-  uint32_t pad;
-} log_rec_t;                                    /* 40 bytes                    */
+  uint16_t pad;                                 /* was u32 when the frame was 30 */
+} log_rec_t;                                    /* 4+32+2+2 = 40 bytes          */
 
 static uint32_t s_head;    /*!< next slot to write, 0..REC_TOTAL-1             */
 static uint32_t s_count;   /*!< valid records                                  */
